Move files to correct package directory
